WIP: progress-tracking test
I've split out a fixtures ns to house the setup/ teardown steps that
will be required in each feature test (i.e. for building cljs, and
managing the server and browser).

I've also split out a helper ns to house functions for exercising the
site via the Taxi API - i.e. for finding and interacting with elements.

This leaves a progress-tracking features file with the actual (pending)
test itself.

The cljs building process may not be ideal but it at least works for
now.