WIP: Fixtures Component
Introduce component into test system to manage database fixtures.

This component takes care of migrating the database, adding-in test data
(at the moment this is just a test user) and then cleaning the database
afterwards. We must use a new component (rather than replacing the
database component with a test version) because the migrations depend
upon the ragtime-component.