Persist progress data
This commit adds `complete-group!` and `group-completed?` database
functions. The truncate command is broadened to include the progress
table as `truncate-database!`.

There are two changes to the base migration:
- user.id and progress.user_id become integers to allow auto
  incrementing behaviour
- progress rows must be unique with respect to user_id and game_id to
  allow updating progress in place

Perhaps these should be part of a new migration rather than changes to
the existing one?

The `with-system` and `test-database` fixtures should be useful for
testing with components.